requirements.txt
bs4
request
urllib3
geopy
wikipedia
pyjokes
tensorflow
numpy
